% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_by_general.R
\name{count_by_general}
\alias{count_by_general}
\title{Generalized Count Function}
\usage{
count_by_general(data, column_names, date = NULL)
}
\arguments{
\item{data}{A data frame.The data should be the raw data frame from which the user wants to extract information.}

\item{column_names}{character vectors containing the names of the columns to group by. It could be one or more
For each specified column, if a value contains a comma, only the part before the comma is kept. So be careful,
if you don't want to split the column values by comma, you should use you own functions. Also only accept comma split.}

\item{date}{A character string containing the name of the date column (optional). If you want to pass a column
which is a date, you should use date to specify the parameter. The date must be in the form of mm dd, yy.}
}
\value{
A tibble containing the count of rows for each group.
}
\description{
This function groups data by one or more categorical variables and returns a summary tibble.
For each specified column, if a value contains a comma, only the part before the comma is kept.
If a date column is specified, only the year is kept.
}
\examples{
\dontrun{
summary_result <- count_by_general(steam_games, c("genre", "language"), date = "release_date")
}
}
